Refactor HvacActuator
